Renaming schema holding folder
