+++ src/server/db_schemas/studentSchema.js
